+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- `NetSparkle.UpdateSystemProfileInformation` is now private
+- `AppCast` no longer takes a Sparkle object and instead takes only those parameters that it needs to operate
+
 ### Removed
 
 ## [0.13.0] - 2017-12-06